+++ libs/ecs/src/def.c
@@ -48,11 +48,11 @@ MAYBE_UNUSED static const EcsCompDef* ecs_def_comp_by_name(const EcsDef* def, co
 EcsDef* ecs_def_create(Allocator* alloc) {
   EcsDef* def = alloc_alloc_t(alloc, EcsDef);
   *def        = (EcsDef){
-             .modules    = dynarray_create_t(alloc, EcsModuleDef, 64),
-             .components = dynarray_create_t(alloc, EcsCompDef, 128),
-             .views      = dynarray_create_t(alloc, EcsViewDef, 128),
-             .systems    = dynarray_create_t(alloc, EcsSystemDef, 128),
-             .alloc      = alloc,
+      .modules    = dynarray_create_t(alloc, EcsModuleDef, 64),
+      .components = dynarray_create_t(alloc, EcsCompDef, 128),
+      .views      = dynarray_create_t(alloc, EcsViewDef, 128),
+      .systems    = dynarray_create_t(alloc, EcsSystemDef, 128),
+      .alloc      = alloc,
   };
   return def;
 }
@@ -205,7 +205,7 @@ void ecs_def_update_order(EcsDef* def, const EcsSystemId system, const i32 order
   ecs_def_system_mutable(def, system)->order = order;
 }
 
-void ecs_def_update_parallel(EcsDef* def, const EcsSystemId system, const u32 parallelCount) {
+void ecs_def_update_parallel(EcsDef* def, const EcsSystemId system, const u16 parallelCount) {
   diag_assert_msg(parallelCount, "Parallel count of 0 is not supported");
   ecs_def_system_mutable(def, system)->parallelCount = parallelCount;
 }
