+++ libs/ecs/src/module.c
@@ -132,6 +132,6 @@ void ecs_module_update_order(EcsModuleBuilder* builder, const EcsSystemId system
 }
 
 void ecs_module_update_parallel(
-    EcsModuleBuilder* builder, const EcsSystemId system, const u32 parallelCount) {
+    EcsModuleBuilder* builder, const EcsSystemId system, const u16 parallelCount) {
   ecs_def_update_parallel(builder->def, system, parallelCount);
 }
